+++ src/index.ts
@@ -19,15 +19,73 @@ const filesLoadedBeforeTrace = Object.keys(require.cache);
 // This file's top-level imports must not transitively depend on modules that
 // do I/O, or continuation-local-storage will not work.
 import * as semver from 'semver';
-import {Config} from './config';
+import {Config, defaultConfig} from './config';
+import * as extend from 'extend';
 import * as path from 'path';
 import * as PluginTypes from './plugin-types';
-import {tracing, Tracing} from './tracing';
-import {Singleton} from './util';
+import {tracing, Tracing, NormalizedConfig} from './tracing';
+import {Singleton, FORCE_NEW, Forceable} from './util';
+import {Constants} from './constants';
 
 export {Config, PluginTypes};
 
 let tracingSingleton: typeof tracing;
+
+/**
+ * Normalizes the user-provided configuration object by adding default values
+ * and overriding with env variables when they are provided.
+ * @param projectConfig The user-provided configuration object. It will not
+ * be modified.
+ * @return A normalized configuration object.
+ */
+function initConfig(projectConfig: Forceable<Config>):
+    Forceable<NormalizedConfig> {
+  // `|| undefined` prevents environmental variables that are empty strings
+  // from overriding values provided in the config object passed to start().
+  const envConfig = {
+    logLevel: Number(process.env.GCLOUD_TRACE_LOGLEVEL) || undefined,
+    projectId: process.env.GCLOUD_PROJECT || undefined,
+    serviceContext: {
+      service:
+          process.env.GAE_SERVICE || process.env.GAE_MODULE_NAME || undefined,
+      version: process.env.GAE_VERSION || process.env.GAE_MODULE_VERSION ||
+          undefined,
+      minorVersion: process.env.GAE_MINOR_VERSION || undefined
+    }
+  };
+
+  let envSetConfig: Config = {};
+  if (!!process.env.GCLOUD_TRACE_CONFIG) {
+    envSetConfig =
+        require(path.resolve(process.env.GCLOUD_TRACE_CONFIG!)) as Config;
+  }
+  // Configuration order of precedence:
+  // 1. Environment Variables
+  // 2. Project Config
+  // 3. Environment Variable Set Configuration File (from GCLOUD_TRACE_CONFIG)
+  // 4. Default Config (as specified in './config')
+  const config = extend(
+      true, {[FORCE_NEW]: projectConfig[FORCE_NEW]}, defaultConfig,
+      envSetConfig, projectConfig, envConfig, {plugins: {}});
+  // The empty plugins object guarantees that plugins is a plain object,
+  // even if it's explicitly specified in the config to be a non-object.
+
+  // Enforce the upper limit for the label value size.
+  if (config.maximumLabelValueSize >
+      Constants.TRACE_SERVICE_LABEL_VALUE_LIMIT) {
+    config.maximumLabelValueSize = Constants.TRACE_SERVICE_LABEL_VALUE_LIMIT;
+  }
+
+  // If the CLS mechanism is set to auto-determined, decide now what it should
+  // be.
+  const ahAvailable = semver.satisfies(process.version, '>=8') &&
+      process.env.GCLOUD_TRACE_NEW_CONTEXT;
+  if (config.clsMechanism === 'auto') {
+    config.clsMechanism = ahAvailable ? 'async-hooks' : 'async-listener';
+  }
+
+  return config;
+}
 
 /**
  * Start the Stackdriver Trace Agent with the given configuration (if provided).
@@ -43,18 +101,15 @@ let tracingSingleton: typeof tracing;
  * trace.start();
  */
 export function start(config?: Config): PluginTypes.TraceAgent {
+  const normalizedConfig = initConfig(config || {});
   // Determine the preferred context propagation mechanism, as
   // continuation-local-storage should be loaded before any modules that do I/O.
-  const ahAvailable = semver.satisfies(process.version, '>=8') &&
-      process.env.GCLOUD_TRACE_NEW_CONTEXT;
-  const agentEnabled = !config || config.enabled !== false;
-  const alAutoPreferred =
-      !ahAvailable && (!config || config.clsMechanism === 'auto');
-  const alUserPreferred = config && (config.clsMechanism === 'async-listener');
-  if (agentEnabled && (alAutoPreferred || alUserPreferred)) {
+  if (normalizedConfig.enabled &&
+      normalizedConfig.clsMechanism === 'async-listener') {
     // This is the earliest we can load continuation-local-storage.
     require('continuation-local-storage');
   }
+
   if (!tracingSingleton) {
     tracingSingleton = require('./tracing').tracing;
   }
@@ -62,7 +117,7 @@ export function start(config?: Config): PluginTypes.TraceAgent {
   try {
     let tracing: Tracing;
     try {
-      tracing = tracingSingleton.create(config || {}, {});
+      tracing = tracingSingleton.create(normalizedConfig, {});
     } catch (e) {
       // An error could be thrown if create() is called multiple times.
       // It's not a helpful error message for the end user, so make it more
